receiver

+++ tests/control_mqtt_receiver.c
@@ -15,6 +15,8 @@
 #include <stdio.h>
 #include <string.h>
 
+#include <sys/time.h>
+
 #define QoS 0
 
 unsigned int total_msg = 500;
@@ -23,7 +25,7 @@ unsigned int started_flag = 0;
 unsigned int stopped_flag = 0;
 
 struct timeval time_start;
-unsigned int time_total = 0;
+double time_total = 0;
 unsigned int count_msg = 0;
